pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "djangophysics"
version = "1.1.0"
description = "Django app to handle and serve physical quanities"
authors = [
    {name = "Frederic Meurou", email = "fm@peabytes.me"},
]
dependencies = [
    "Django>=4.1",
    "django-cors-headers>=4.3.0",
    "django-createsuperuser",
    "django-extensions>=3.2",
    "django-filter>=22.1",
    "django-sendfile>=0.3",
    "djangorestframework>=3.14",
    "drf-yasg>=1.21",
    "markdown>=3.4",
    "lxml~=4.9",
    "django-redis>=5.2.0",
    "pysendfile~=2.0",
    "gunicorn",
    "pycountry>=22",
    "countryinfo>=0.1",
    "timezonefinder>=6.1",
    "iso4217>=1.11",
    "forex-python>=1.0",
    "Babel>=2.10",
    "Pint>=0.19",
    "networkx~=2.8",
    "sympy~=1.10",
    "requests~=2.28",
    "channels>=4.0",
    "uncertainties>=3.1",
    "ariadne>=0.16",
    "ariadne-django>=0.3.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "coverage>=5.3.1",
    "pylint-venv>=2.1.1",
    "django-extensions>=3.2",
    "ipython>=7.23.1"
]

[project.optional-dependencies]
postgres = [
    "psycopg-binary",
    "daphne"
]
mysql = [
    "mysqlclient>=2.0.3",
    "daphne"
]
